Add files via upload
Here are the Files I have been working on 2/25-2/28.  They include major modifications to Game_Level, Game_Board, AI and Ship classes.  The main file is configured to run the Catch2 framework tests, not to generate a BattleshipBoard.  There is a fatal error in the Board class about deleting memory for the _Board array or something like that.  The _Board is not updating itself and therefore the ship_Overlap funcition is not working.